+++ src/diagram/components/icons/icon-color-menu.tsx
