+++ src/diagram/components/icons/icon-expand.tsx
