+++ src/diagram/components/icons/icon-warning.tsx
